+++ src/ffpack/lcc/astmCounting.py
@@ -58,7 +58,7 @@ def astmLevelCrossingCounting( data, refLevel=0.0, levels=None, aggregate=True )
         levels = np.array( sorted( set( levels ) ) )
 
     # Remove the intermediate value first
-    data = np.array( generalUtils.sequencePeakAndValleys( data, keepEnds=True ) )
+    data = np.array( generalUtils.sequencePeakValleyFilter( data, keepEnds=True ) )
 
     rstDict = defaultdict( int )
     rstSeq = [ ]
@@ -182,7 +182,7 @@ def astmSimpleRangeCounting( data, aggregate=True ):
         raise ValueError( "Input data length should be at least 2")
 
     # Remove the intermediate value first
-    data = np.array( generalUtils.sequencePeakAndValleys( data, keepEnds=True ) )
+    data = np.array( generalUtils.sequencePeakValleyFilter( data, keepEnds=True ) )
 
     rstDict = defaultdict( int )
     rstSeq = [ ]
@@ -237,7 +237,7 @@ def astmRainflowCounting( data, aggregate=True ):
         raise ValueError( "Input data length should be at least 2")
 
     # Remove the intermediate value first
-    data = np.array( generalUtils.sequencePeakAndValleys( data, keepEnds=True ) )
+    data = np.array( generalUtils.sequencePeakValleyFilter( data, keepEnds=True ) )
 
     dequeA = deque()
     dequeB = deque( [ i for i in data ] )
@@ -330,7 +330,7 @@ def astmRangePairCounting( data, aggregate=True ):
         raise ValueError( "Input data length should be at least 2")
 
     # Remove the intermediate value first
-    data = np.array( generalUtils.sequencePeakAndValleys( data, keepEnds=True ) )
+    data = np.array( generalUtils.sequencePeakValleyFilter( data, keepEnds=True ) )
     indices = np.array( range( -1, len( data ) - 1 ) )
 
     def checkPreviousThree( indices, i ):
@@ -432,7 +432,7 @@ def astmRainflowRepeatHistoryCounting( data, aggregate=True ):
         raise ValueError( "Input data should be repeating")
 
     # Remove the intermediate value first
-    data = np.array( generalUtils.sequencePeakAndValleys( data, keepEnds=True ) )
+    data = np.array( generalUtils.sequencePeakValleyFilter( data, keepEnds=True ) )
     # search the peak and shift the data
     index = data.argmax( axis=0 )
     n = len( data )
@@ -441,7 +441,7 @@ def astmRainflowRepeatHistoryCounting( data, aggregate=True ):
             data[ i ] = data[ i + 1 ]
         data = np.roll( data, -index )
         # need to remove the intermediate value again
-        data = np.array( generalUtils.sequencePeakAndValleys( data, keepEnds=True ) )
+        data = np.array( generalUtils.sequencePeakValleyFilter( data, keepEnds=True ) )
 
     indices = np.array( range( -1, len( data ) - 1 ) )