+++ tests/utils/test_utils_generalUtils.py
@@ -6,209 +6,209 @@ import pytest
 
 
 ###############################################################################
-# Test sequencePeakAndValleys
+# Test sequencePeakValleyFilter
 ###############################################################################
-def test_sequencePeakAndValleys_noPoints_valueError():
+def test_sequencePeakValleyFilter_noPoints_valueError():
     data = [ ]
     with pytest.raises( ValueError ):
-        _ = utils.sequencePeakAndValleys( data )
+        _ = utils.sequencePeakValleyFilter( data )
     
     data = [ ]
     with pytest.raises( ValueError ):
-        _ = utils.sequencePeakAndValleys( data, keepEnds=True )
+        _ = utils.sequencePeakValleyFilter( data, keepEnds=True )
 
 
-def test_sequencePeakAndValleys_twoDimData_valueError():
+def test_sequencePeakValleyFilter_twoDimData_valueError():
     data = [ [ 1.0, 2.0 ], [ 3.0, 4.0 ] ]
     with pytest.raises( ValueError ):
-        _ = utils.sequencePeakAndValleys( data )
+        _ = utils.sequencePeakValleyFilter( data )
     
     data = [ [ 1.0, 2.0 ], [ 3.0, 4.0 ] ]
     with pytest.raises( ValueError ):
-        _ = utils.sequencePeakAndValleys( data, keepEnds=True )
+        _ = utils.sequencePeakValleyFilter( data, keepEnds=True )
 
 
-def test_sequencePeakAndValleys_onePointsKeepEnds_valueError():
+def test_sequencePeakValleyFilter_onePointsKeepEnds_valueError():
     data = [ 1.0 ]
     with pytest.raises( ValueError ):
-        _ = utils.sequencePeakAndValleys( data )
+        _ = utils.sequencePeakValleyFilter( data )
     
     data = [ 1.0 ]
     with pytest.raises( ValueError ):
-        _ = utils.sequencePeakAndValleys( data, keepEnds=True )
+        _ = utils.sequencePeakValleyFilter( data, keepEnds=True )
 
 
-def test_sequencePeakAndValleys_twoPointsNotKeepEnds_valueError():
+def test_sequencePeakValleyFilter_twoPointsNotKeepEnds_valueError():
     data = [ 1.0, 2.0 ]
     with pytest.raises( ValueError ):
-        _ = utils.sequencePeakAndValleys( data )
+        _ = utils.sequencePeakValleyFilter( data )
 
 
-def test_sequencePeakAndValleys_twoPointsKeepEnds_pointskept():
+def test_sequencePeakValleyFilter_twoPointsKeepEnds_pointskept():
     data = [ -0.5, 1.0 ]
     # Keep ends
-    calRst = utils.sequencePeakAndValleys( data, keepEnds=True )
+    calRst = utils.sequencePeakValleyFilter( data, keepEnds=True )
     expectedRst = [ -0.5, 1.0 ]
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-def test_sequencePeakAndValleys_threePointsSimple_pass():
+def test_sequencePeakValleyFilter_threePointsSimple_pass():
     data = [ -0.5, 1.0, 0.0 ]
 
     # case 1: do not keep ends
-    calRst = utils.sequencePeakAndValleys( data )
+    calRst = utils.sequencePeakValleyFilter( data )
     expectedRst = [ 1.0 ]
     np.testing.assert_allclose( calRst, expectedRst )
 
     # case 2: keep ends
-    calRst = utils.sequencePeakAndValleys( data, keepEnds=True )
+    calRst = utils.sequencePeakValleyFilter( data, keepEnds=True )
     expectedRst = [ -0.5, 1.0, 0.0 ]
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-def test_sequencePeakAndValleys_threePointsWithSameValue_depends():
+def test_sequencePeakValleyFilter_threePointsWithSameValue_depends():
     # case 1: last two the same
     data = [ -0.5, 1.0, 1.0 ]
     
-    calRst = utils.sequencePeakAndValleys( data )
+    calRst = utils.sequencePeakValleyFilter( data )
     expectedRst = [  ]
     np.testing.assert_allclose( calRst, expectedRst )
 
-    calRst = utils.sequencePeakAndValleys( data, keepEnds=True )
+    calRst = utils.sequencePeakValleyFilter( data, keepEnds=True )
     expectedRst = [ -0.5, 1.0 ]
     np.testing.assert_allclose( calRst, expectedRst )
 
     # case 2: first two the same
     data = [ 1.0, 1.0, 2.0 ]
     
-    calRst = utils.sequencePeakAndValleys( data )
+    calRst = utils.sequencePeakValleyFilter( data )
     expectedRst = [  ]
     np.testing.assert_allclose( calRst, expectedRst )
 
-    calRst = utils.sequencePeakAndValleys( data, keepEnds=True )
+    calRst = utils.sequencePeakValleyFilter( data, keepEnds=True )
     expectedRst = [ 1.0, 2.0 ]
     np.testing.assert_allclose( calRst, expectedRst )
 
     # case 3: three values the same
     data = [ 1.0, 1.0, 1.0 ]
-    calRst = utils.sequencePeakAndValleys( data )
+    calRst = utils.sequencePeakValleyFilter( data )
     expectedRst = [  ]
     np.testing.assert_allclose( calRst, expectedRst )
 
-    calRst = utils.sequencePeakAndValleys( data, keepEnds=True )
+    calRst = utils.sequencePeakValleyFilter( data, keepEnds=True )
     expectedRst = [ 1.0, 1.0 ]
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-def test_sequencePeakAndValleys_fourPointsWithSameValue_depends():
+def test_sequencePeakValleyFilter_fourPointsWithSameValue_depends():
     # case 1: last three the same
     data = [ -0.5, 1.0, 1.0, 1.0 ]
     
-    calRst = utils.sequencePeakAndValleys( data )
+    calRst = utils.sequencePeakValleyFilter( data )
     expectedRst = [  ]
     np.testing.assert_allclose( calRst, expectedRst )
 
-    calRst = utils.sequencePeakAndValleys( data, keepEnds=True )
+    calRst = utils.sequencePeakValleyFilter( data, keepEnds=True )
     expectedRst = [ -0.5, 1.0 ]
     np.testing.assert_allclose( calRst, expectedRst )
 
     # case 2: first three the same
     data = [ 1.0, 1.0, 1.0, 2.0 ]
     
-    calRst = utils.sequencePeakAndValleys( data )
+    calRst = utils.sequencePeakValleyFilter( data )
     expectedRst = [  ]
     np.testing.assert_allclose( calRst, expectedRst )
 
-    calRst = utils.sequencePeakAndValleys( data, keepEnds=True )
+    calRst = utils.sequencePeakValleyFilter( data, keepEnds=True )
     expectedRst = [ 1.0, 2.0 ]
     np.testing.assert_allclose( calRst, expectedRst )
 
     # case 3: first two the same, last two the same
     data = [ 1.0, 1.0, 2.0, 2.0 ]
     
-    calRst = utils.sequencePeakAndValleys( data )
+    calRst = utils.sequencePeakValleyFilter( data )
     expectedRst = [  ]
     np.testing.assert_allclose( calRst, expectedRst )
 
-    calRst = utils.sequencePeakAndValleys( data, keepEnds=True )
+    calRst = utils.sequencePeakValleyFilter( data, keepEnds=True )
     expectedRst = [ 1.0, 2.0 ]
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-def test_sequencePeakAndValleys_fivePointsWithSameValue_depends():
+def test_sequencePeakValleyFilter_fivePointsWithSameValue_depends():
     # case 1: three in the middle the same
     data = [ -0.5, 1.0, 1.0, 1.0, 0.0 ]
     
-    calRst = utils.sequencePeakAndValleys( data )
+    calRst = utils.sequencePeakValleyFilter( data )
     expectedRst = [ 1.0 ]
     np.testing.assert_allclose( calRst, expectedRst )
 
-    calRst = utils.sequencePeakAndValleys( data, keepEnds=True )
+    calRst = utils.sequencePeakValleyFilter( data, keepEnds=True )
     expectedRst = [ -0.5, 1.0, 0.0 ]
     np.testing.assert_allclose( calRst, expectedRst )
 
     # case 2: last three the same
     data = [ -0.5, 1.0, 2.0, 2.0, 2.0 ]
     
-    calRst = utils.sequencePeakAndValleys( data )
+    calRst = utils.sequencePeakValleyFilter( data )
     expectedRst = [  ]
     np.testing.assert_allclose( calRst, expectedRst )
 
-    calRst = utils.sequencePeakAndValleys( data, keepEnds=True )
+    calRst = utils.sequencePeakValleyFilter( data, keepEnds=True )
     expectedRst = [ -0.5, 2.0 ]
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-def test_sequencePeakAndValleys_normalUseOnlyPeakAndValleys_pass():
+def test_sequencePeakValleyFilter_normalUseOnlyPeakAndValleys_pass():
     data = [ -0.5, 1.0, -2.0, 3.0, -1.0, 4.5, -2.5, 3.5, -1.5, 1.0 ]
     # Do not keep ends
-    calRst = utils.sequencePeakAndValleys( data )
+    calRst = utils.sequencePeakValleyFilter( data )
     expectedRst = [ 1.0, -2.0, 3.0, -1.0, 4.5, -2.5, 3.5, -1.5 ]
     np.testing.assert_allclose( calRst, expectedRst )
 
     data = [ -0.5, 1.0, -2.0, 3.0, -1.0, 4.5, -2.5, 3.5, -1.5, 1.0 ]
     # Keep ends
-    calRst = utils.sequencePeakAndValleys( data, keepEnds=True )
+    calRst = utils.sequencePeakValleyFilter( data, keepEnds=True )
     expectedRst = [ -0.5, 1.0, -2.0, 3.0, -1.0, 4.5, -2.5, 3.5, -1.5, 1.0 ]
     np.testing.assert_allclose( calRst, expectedRst )
 
     data = [ 1.0, 2.0, 0.5, 3.0, 1.0, 4.5, 2.5, 3.5, 1.5, 4.0 ]
     # Do not keep ends
-    calRst = utils.sequencePeakAndValleys( data )
+    calRst = utils.sequencePeakValleyFilter( data )
     expectedRst = [ 2.0, 0.5, 3.0, 1.0, 4.5, 2.5, 3.5, 1.5 ]
     np.testing.assert_allclose( calRst, expectedRst )
 
     data = [ 1.0, 2.0, 0.5, 3.0, 1.0, 4.5, 2.5, 3.5, 1.5, 4.0 ]
     # Keep ends
-    calRst = utils.sequencePeakAndValleys( data, keepEnds=True )
+    calRst = utils.sequencePeakValleyFilter( data, keepEnds=True )
     expectedRst = [ 1.0, 2.0, 0.5, 3.0, 1.0, 4.5, 2.5, 3.5, 1.5, 4.0 ]
     np.testing.assert_allclose( calRst, expectedRst )
 
     data = [ -1.0, -2.0, -0.5, -3.0, -1.0, -4.5, -2.5, -3.5, -1.5, -4.0 ]
     # Do not keep ends
-    calRst = utils.sequencePeakAndValleys( data )
+    calRst = utils.sequencePeakValleyFilter( data )
     expectedRst = [ -2.0, -0.5, -3.0, -1.0, -4.5, -2.5, -3.5, -1.5 ]
     np.testing.assert_allclose( calRst, expectedRst )
 
     data = [ -1.0, -2.0, -0.5, -3.0, -1.0, -4.5, -2.5, -3.5, -1.5, -4.0 ]
     # Keep ends
-    calRst = utils.sequencePeakAndValleys( data, keepEnds=True )
+    calRst = utils.sequencePeakValleyFilter( data, keepEnds=True )
     expectedRst = [ -1.0, -2.0, -0.5, -3.0, -1.0, -4.5, -2.5, -3.5, -1.5, -4.0 ]
     np.testing.assert_allclose( calRst, expectedRst )
 
 
-def test_sequencePeakAndValleys_normalUseExtraPointsInSequence_pass():
+def test_sequencePeakValleyFilter_normalUseExtraPointsInSequence_pass():
     data = [ -0.5, 0.0, 1.0, -1.0, -2.0, -1.0, 1.5, 3.0, 2.5, -1.0, 0.5, 1.5, 4.5, 
              3.5, 1.0, -1.0, -2.5, -1.5, 3.0, 3.5, 1.5, 0.0, -1.5, 0.5, 1.0 ]
     # Do not keep ends
-    calRst = utils.sequencePeakAndValleys( data )
+    calRst = utils.sequencePeakValleyFilter( data )
     expectedRst = [ 1.0, -2.0, 3.0, -1.0, 4.5, -2.5, 3.5, -1.5 ]
     np.testing.assert_allclose( calRst, expectedRst )
 
     data = [ -0.5, 0.0, 1.0, -1.0, -2.0, -1.0, 1.5, 3.0, 2.5, -1.0, 0.5, 1.5, 4.5, 
              3.5, 1.0, -1.0, -2.5, -1.5, 3.0, 3.5, 1.5, 0.0, -1.5, 0.5, 1.0 ]
     # Keep ends
-    calRst = utils.sequencePeakAndValleys( data, keepEnds=True )
+    calRst = utils.sequencePeakValleyFilter( data, keepEnds=True )
     expectedRst = [ -0.5, 1.0, -2.0, 3.0, -1.0, 4.5, -2.5, 3.5, -1.5, 1.0 ]
     np.testing.assert_allclose( calRst, expectedRst )